Structured Refactor and XML Documentation
- Refactor geometry code into focused helpers
- Remove unused TemporaryObjectManager
- Refactor Axys commands
- Reorganize managers into subfolders
- Add comprehensive XML documentation
